LPS-118440 Use the option label instead of the option ID when the ID is generated randomly
