+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/java/com/liferay/dynamic/data/mapping/form/field/type/internal/checkbox/multiple/CheckboxMultipleDDMFormFieldValueAccessor.java
@@ -15,6 +15,9 @@
 package com.liferay.dynamic.data.mapping.form.field.type.internal.checkbox.multiple;
 
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldValueAccessor;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
+import com.liferay.dynamic.data.mapping.model.LocalizedValue;
 import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
 import com.liferay.petra.string.CharPool;
@@ -25,10 +28,13 @@ import com.liferay.portal.kernel.json.JSONException;
 import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.util.Locale;
 import java.util.function.IntFunction;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -55,13 +61,24 @@ public class CheckboxMultipleDDMFormFieldValueAccessor
 	public JSONArray getValue(
 		DDMFormFieldValue ddmFormFieldValue, Locale locale) {
 
-		Value value = ddmFormFieldValue.getValue();
+		JSONArray optionsValuesJSONArray = getOptionsValuesJSONArray(
+			ddmFormFieldValue, locale);
 
-		if (value == null) {
-			return createJSONArray("[]");
+		Matcher matcher = _pattern.matcher(optionsValuesJSONArray.getString(0));
+
+		if (!matcher.matches() ||
+			(ddmFormFieldValue.getDDMFormValues() == null)) {
+
+			return optionsValuesJSONArray;
 		}
 
-		return createJSONArray(value.getString(locale));
+		StringBundler sb = new StringBundler(3);
+
+		sb.append(StringPool.OPEN_BRACKET);
+		sb.append(getOptionsLabels(ddmFormFieldValue, locale));
+		sb.append(StringPool.CLOSE_BRACKET);
+
+		return createJSONArray(sb.toString());
 	}
 
 	@Override
@@ -118,10 +135,72 @@ public class CheckboxMultipleDDMFormFieldValueAccessor
 		}
 	}
 
+	protected DDMFormFieldOptions getDDMFormFieldOptions(
+		DDMFormFieldValue ddmFormFieldValue) {
+
+		DDMFormField ddmFormField = ddmFormFieldValue.getDDMFormField();
+
+		return ddmFormField.getDDMFormFieldOptions();
+	}
+
+	protected String getOptionsLabels(
+		DDMFormFieldValue ddmFormFieldValue, Locale locale) {
+
+		JSONArray optionsValuesJSONArray = getOptionsValuesJSONArray(
+			ddmFormFieldValue, locale);
+
+		if (optionsValuesJSONArray.length() == 0) {
+			return StringPool.BLANK;
+		}
+
+		DDMFormFieldOptions ddmFormFieldOptions = getDDMFormFieldOptions(
+			ddmFormFieldValue);
+
+		StringBundler sb = new StringBundler(
+			(optionsValuesJSONArray.length() * 2) - 1);
+
+		for (int i = 0; i < optionsValuesJSONArray.length(); i++) {
+			String optionValue = optionsValuesJSONArray.getString(i);
+
+			LocalizedValue optionLabel = ddmFormFieldOptions.getOptionLabels(
+				optionValue);
+
+			if (optionLabel != null) {
+				sb.append(HtmlUtil.escape(optionLabel.getString(locale)));
+			}
+			else {
+				sb.append(optionValue);
+			}
+
+			sb.append(StringPool.COMMA_AND_SPACE);
+		}
+
+		if (sb.index() > 0) {
+			sb.setIndex(sb.index() - 1);
+		}
+
+		return sb.toString();
+	}
+
+	protected JSONArray getOptionsValuesJSONArray(
+		DDMFormFieldValue ddmFormFieldValue, Locale locale) {
+
+		Value value = ddmFormFieldValue.getValue();
+
+		if (value == null) {
+			return createJSONArray("[]");
+		}
+
+		return createJSONArray(value.getString(locale));
+	}
+
 	@Reference
 	protected JSONFactory jsonFactory;
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		CheckboxMultipleDDMFormFieldValueAccessor.class);
+
+	private static final Pattern _pattern = Pattern.compile(
+		"^[A-Za-z]*[0-9]{8}$");
 
 }
