+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-form-field-type/src/main/java/com/liferay/dynamic/data/mapping/form/field/type/internal/radio/RadioDDMFormFieldValueAccessor.java
@@ -15,11 +15,17 @@
 package com.liferay.dynamic.data.mapping.form.field.type.internal.radio;
 
 import com.liferay.dynamic.data.mapping.form.field.type.DDMFormFieldValueAccessor;
+import com.liferay.dynamic.data.mapping.model.DDMFormField;
+import com.liferay.dynamic.data.mapping.model.DDMFormFieldOptions;
+import com.liferay.dynamic.data.mapping.model.LocalizedValue;
 import com.liferay.dynamic.data.mapping.model.Value;
 import com.liferay.dynamic.data.mapping.storage.DDMFormFieldValue;
+import com.liferay.portal.kernel.util.HtmlUtil;
 
 import java.util.Locale;
 import java.util.function.IntFunction;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -42,9 +48,54 @@ public class RadioDDMFormFieldValueAccessor
 
 	@Override
 	public String getValue(DDMFormFieldValue ddmFormFieldValue, Locale locale) {
+		String optionValue = getOptionValue(ddmFormFieldValue, locale);
+
+		Matcher matcher = _pattern.matcher(optionValue);
+
+		if (!matcher.matches() ||
+			(ddmFormFieldValue.getDDMFormValues() == null)) {
+
+			return optionValue;
+		}
+
+		return getOptionLabel(ddmFormFieldValue, locale);
+	}
+
+	protected DDMFormFieldOptions getDDMFormFieldOptions(
+		DDMFormFieldValue ddmFormFieldValue) {
+
+		DDMFormField ddmFormField = ddmFormFieldValue.getDDMFormField();
+
+		return ddmFormField.getDDMFormFieldOptions();
+	}
+
+	protected String getOptionLabel(
+		DDMFormFieldValue ddmFormFieldValue, Locale locale) {
+
+		String optionValue = getOptionValue(ddmFormFieldValue, locale);
+
+		DDMFormFieldOptions ddmFormFieldOptions = getDDMFormFieldOptions(
+			ddmFormFieldValue);
+
+		LocalizedValue optionLabel = ddmFormFieldOptions.getOptionLabels(
+			optionValue);
+
+		if (optionLabel == null) {
+			return optionValue;
+		}
+
+		return HtmlUtil.escape(optionLabel.getString(locale));
+	}
+
+	protected String getOptionValue(
+		DDMFormFieldValue ddmFormFieldValue, Locale locale) {
+
 		Value value = ddmFormFieldValue.getValue();
 
 		return value.getString(locale);
 	}
+
+	private static final Pattern _pattern = Pattern.compile(
+		"^[A-Za-z]*[0-9]{8}$");
 
 }
